feat: Rename not-only-case (task7-sprint1-model)
