+++ src/main/java/model/mongohelperss/MongoAttributes.java
@@ -1,4 +1,4 @@
-package model.mongohelpers;
+package model.mongohelperss;
 
 public class MongoAttributes {
     public final static String IdAttribute = "_id";
